OCI containers with Podman and podman-systemd
It sort-of works, but needs MUCH more work to be even slightly reliable.
In particular, the invisible dependency between apt.Package and oci.* is
annoying; things fail on first deploy and succeed on subsequent ones.
Perhaps host.Host could have 'onboarding' steps - but the naive way of
doing this creates a circular dependency...