Flock, a mesh networking system based on Nebula
This is unfinished. Already working:
* CA and certificate management
* Nebula lighthouses
* Lighthouse DNS
* TUN/TAP in sidecar container

Pod networking does NOT work. Having the pod network as bridge means that TUN
device created by sidecar is not accessible by other containers. To fix,
containers should reuse sidecar's network stack with --network container:<name>.

The downside of this is that if we place containers on same network.
Kubernetes does it this way, so it is probably fine! Just need to refactor
code to support this without too much spaghettification...