Add support for HA Postgres deployments with Patroni
No support for configuring anything, terrible automated tests,
hacked-together container image on top of Ubuntu...

But it works. Brought a VM down, and failover happened exactly as it should.
Brought it back up, and the old primary was now a follower.

Next up: implement that service DNS, so this can actually be used!